Added ProxyPlugin
ProxyPlugin forwards data received from OCC to a network socket,
if client is connected. Most of the code was similar to AdaraPlugin,
so an abstract BaseSocketPlugin was created. ProxyPlugin was used
for initial performance tests, where the client was a NetCat program
dumping all data /dev/null.
Forwarding command packets from client to OCC is to be implemented
when needed.