Let locally copied buffer not stall but rather delegate that event to OCC API.
When the local buffer was getting full, the copy thread would quit
and a stall signal would be sent to consumers. This imposed two possible ways
to stall, one in the OCC hw and one in software.
With the patch in place, the copy thread now checks to see if there's place and
it gives more time to consumers otherwise. It never stalls. Instead it pushes
stalling back to OCC. If OCC has room all is fine, otherwise OCC stalls which
is handled already with the status checking.
The patch is necessary for a piped OCC whichs is feed the data from a file,
obviously very fast. It also reduces redundant information visible to the user.