Added workaround for v5.4 module
Vision is full of 5.4 ROCs which can't all be upgraded ATM. The workaround
modified the incoming packet for its length. Since packets must not be
modified in place as the memory is shared between all plugins and also
the port driver, we must make a copy. Modified packet is then
sent to the parses function.