Initial version of calibration plugin
This commit introduces a functional but incomplete detector calibration
plugin for nED.  It creates an EPICS v4 record named "raw_cal_data" and
parses incoming data packets to update the record.  It needs to be able
to differentiate between normal data and "raw data", the latter of which
is used for calibration.  This is planned to be done via an additional
process variable.