Rework Calibration plugin to work with BasePvaPlugin
BasePvaPlugin (along with PvaNeutronData) class was created to abstract
many of the features of the original CalibrationPlugin work, such as
the creation of a common EPICS V4 neutron data record, a method for
plugins to differentiate different types of incoming detector data (raw,
normal, extended), and additional features for updating EPICS V4 data
records.

This patch removes CalibrationPlugin in favor of CalibrationPvaPlugin.
CalibrationPvaPlugin inherits from BasePvaPlugin, and is therefore
greatly simplified.