Updated RocPvaPlugin to conform to new C-style callbacks
Klemen and I determined that we needed to move much of the standard
event processing code into the BasePvaPlugin class, in order to
avoid massive code duplication.  Now, the base class processes the
DasPacketList and performs a C-callback to the specialized class in
order to process the data payload.  There is also a callback to post
the data to the EPICS V4 record.  This commit updated RocPvaPlugin
to conform to this improved model.  We also expect performance
improvements because we now be able to update only the EPICS fields
which have actually changed.