Detector configuration parameters get read-back capability
Previously the READ_CONFIG command to a detector did not have any effect
on configuration EPICS PVs because asyn device support is uni-directional.
With a trick (adding 3 more records for every config record) we're now
able to also read values from nED.
This is useful with READ_CONFIG command to show what the configuration
in detector really is. Or to reset any changes made that are not yet
pushed to detector. It also allows seamless migration from dcomserver.

It proved to be a valuable tool during calibration development and there
at least some interest to see use that in the future.