bugfix: Handle non 4-byte aligned status and config packets.
LVDS based modules can produce non 4-byte aligned command responses but
DSP rounds them up 4-byte aligned, filling missing bytes with 0. The DAS
packet that we receive from OCC is thus 4-byte aligned, but the actual
content is not and we need to discard padding bytes.
Since all responses have static size and we know what it should be, this
is easy. Only the length checks need to be aware of the alignment.