Account for 2-byte alignment for outgoing LVDS packets
When DSP is passing through responses from submodules, it "corrects"
2-byte aligned responses to 4 bytes. OCC protocol is 4-byte aligned.
But than the length in the packet header does not match the actual
responses from submodule and software needs to be aware of that.
When software is sending the same data out to the sub-module,
it must send 4-byte aligned block of data (because of the protocol,
everybody expects 4-byte alignment), but the length in the header
must reflect the actual payload size which might be 2-byte aligned.
Otherwise the submodule refuses the command saying it's the length
error. This breaks the OCC protocol, but I guess the DSP doesn't mind.
For now it's only the DSP that receives and processes that data.