Added support for LPSD ROC v57.
I have added the necessary support for ROC v57, which was derived
mostly from v54.  There are two new commands, CMD_WRITE_CONFIG_EX (0xE0)
and CMD_READ_CONFIG_EX (0xE1).  These correspond to a new
"extended" configuration address space in the ROC.  Still need to
update to ROC OPI screens to be able to send these commands.  There
is also a new command to reset status counters (0x25).

Support has been added for Test Pattern generation registers, which
are the new residents of this extended configuration space.

The extended configuration address space was added in the name of
backwards compatibility with dComServer.  In this way, new parameters
can be added (in this case, Test Pattern generation registers) without
affecting the size or layout of the existing configuration space.