Converted signed IntRelease to unsigned value.
As discussed with Bogdan it's not worth the effort (at least not at this moment)
to add support for signed values. As implemented today, it works in direction from
nED to ROCs, but not in the other direction. Value of -6 gets converted to 65530
and sent to ROC, but when read back the value remains 65530 as nED considers
all registers as unsigned.
For IntRelease it's just user convinience to represent it as negative value and
most beamlines don't change that ever. So 65530 is just as good.