Enable heartbeats when there's no neutron data but there are RTDL packets
It was observed that when DSP is not acquiring data the SMS is building up a table
of pulses it's waiting to receive data for. The proper solution would be that AdaraPlugin
sends an empty data packet when there's no data for that pulse, but since RTDL packets
and data packets are not time aligned (RTDL usually comes couple of pulses ahead),
this would require buffering RTDL in nED.
More elegant solution for nED is to send heartbeat when there is no neutron or
meta data even though RTDL is coming through.