Updated Roc v52 and v56 register maps
Based on input from Vlad, fixed errors in ROC register maps in the
v52 and v56 versions.  These versions have identical config
registers, except for the addition of test pattern generation
registers in v56.  The changes made were to interleave the Scale
and Offset config registers, and modify the register widths of
a few registers in Section F.