Expanded OCC rates, fixed some minor issues
Chasing the corrupted data problem, several improvements to the code were made.
In addition to the processing rate previously displayed on screen, also the OCC
receive rate (most left), copy from DMA to buffer rate (middle), processing rate
(most right as before) are now displayed. This allows us to easily track problems
where the receive side is accepting data, but the processing takes long time
which results in processing speed to be 0 for most of the time and jump to a
high number every now and then.

Rate calculation has been much simplified in the processing thread. It's now
done in the CirculaBuffer implementations. It's not I/O intr any more,
data is collected internally and polled based on record SCAN specs (1 second).
The calculated rate is average since last read.
Except for receive rate, where hardware calculates it and is polled every 5s
when refreshing status.

Minor improvement to copy thread, not to get into endless loop when buffer is
full. Another improvement to be able to get out of error state when copy stalled
(m_error was not being properly cleared).

Finally, typo in FpgaTempAlarm record was causing the corrupted data problem.
When DMA stalled, this guy would re-enable the DMA every 60 seconds, effectively
clearing contents of DMA buffer, but not synchronizing it with copy buffer. Copy
buffer would then continue as if nothing happened, but sometimes there was a
partial packet sitting there which nED detected as corrupt data.