Add FlatField transformation plugin.
This is a work in progress commit. So far plugin can read in X and Y
correction tables for single position when instantiated. It will apply
X,Y->pixel transformation for every neutron event that belongs to loaded
position. It will do so in parallelized-safe way, although not (yet)
parallelized.

Missing functionality includes:
* load multiple positions
* load and apply photo-sum correction