Flush PV record when switching data modes
Data from previous data mode must be flushed out if it's not being updated
in the new mode, or it will get posted for every PV record change for the new
data mode, even though it hasn't changed. User would probably not recognize
it.
Data is flushed whenever there's a data mode change or when Flat-Fielded
data is enabled in AcpcPvaPlugin. All fields are flushed except for timeStamp,
where the time remains the same but the userTag increments by one.