Fixes for Pulse commands of ADCROC
Previous commits failed to add code to BaseModulePlugin to
support the Pulse and Pulse clear commands.  I have also updated
the database templates for BaseModulePlugin and the DebugPlugin
accordingly.

The pulse and pulse clear commands were tested in the lab using
oscilloscope, and found to be working correctly.  We also tested
data acquisition in fake trigger mode, and data rate controls.
Everything appears to work and is ready for field integration with
actual pulse magnet.