Changed test pattern registers for ROC 5.6 and 5.7 only.
As discussed today with Vlad and Bogdan we'll leave ROC 5.6 and ROC 5.7
with bit 14 as alternate test pattern generator, in addition to bit 15
as a global test pattern switch. This should be corrected in newer
ROC versions, but since deployed on VISION, CORELLI and USANS we'll
update the software for now.
When enabled, ramp test pattern generator will override any other
data produced by ROC including alternate test pattern generator which
makes this solution work for the test pattern analyzer.