Add support for ROC 4.7
There are 2 different ROC firmware both identified with version 5.2. The
difference is in the hardware they run on. So far we supported new hardware
(hw=5.2) and currently all firmwares 5.x are based on that. It has recently
been discovered that hw=2.0,fw=5.2 also exists but has very different
register interface, much like the rest of 4.x firmwares. nED distinguishes
detectors solely on firmware version (aka interface version), but not hardware
version.

None of the 4.x series we have sources for, only hw=2.0,fw=5.2 is under Vlad's
control. But we can't use the 5.2 version unless we modify software policy.
For now the idea is to re-version hw=2.0,fw=5.2 to something else in 4.x range.

I'm picking 4.7 for now (no reference in dcomserver of such version) and we can
change it later.