Revert commit with signed values, will do proper support in August.
The support was only partial. It treated all registers as signed
two-complement but some of the registers are trully unsigned. Also,
as it turns out some of the registers in the same firmware can also
be signed magnitude. Maybe there's also one complement register
somewhere out there.

With signed magnitude the number of bits must be exacytly speicified.
Less of a problem for two complement. But it has been observed that
some registers might be specified wrongly. They will have to be
revised when turned into signed. Because of that the change is not
going mainstream for now and will have to do proper verification
after summer shutdown.