Merged OCC status into single PV and added PV to disable RX on any error.
The PortStatus PV now lists all possible errors rather than having two
separate PVs. When an error is detected, the db logic switches off received
part of OCC through the RxEn PV. The RxEn PV behavior change not to reset OCC
when it's turned off but rather when it's turned on. This retains the last
error in PortStatus, but still clears any errors before starting up.