Split 'Corruputed data' message in more detailed errors
There are only 2 data-validity checks in nED, both are based on data length:
* there must be enough data in buffer to parse the packet
* packet length must not exceed hardcoded limit of 3600B

When detected, either of those two produced the same error message.
In SEQUOIA runaway tests the first one was observed for the first time and it
was not obvious right away. Only after digging in logs, the real reason was revealed.
It's better to specify exact reason even on PV level.