bugfix: Segfault in PixelMapPlugin
Invalid packet header caused misinterpretation of the packet. The 3rd dword
in the packet (aka cmdinfo) tells the type of the data and, in case of
neutron/meta data, also flags whether RTDL subheader is present. Mis-formed
cmdinfo caused nED to interpret 8 byte payload packet as valid neutron data
with 24 byte of RTDL header part of it. The payload obviously did not have
enough room, causing miscalculation of number of events in the packet. The
calculated number was negative and the number of events was huge. When
PixelMapPlugin was transforming the events it went over the buffer boundaries.

The patch adds extra check for packet length when RTDL header is told to
be there.

It was first observed on SEQUOIA but not inspected. It was found, debugged
and fixed on HYSPEC and proven to work well on SEQUOIA.