Smarter Acquiring state handling
Drive Acquiring alarm based on the start/stop request. Almost all modules provide an Acquiring
register that is updated every time READ_STATUS is executed and returned. It's an error when
module gets started/stopped and the Acquiring
register doesn't agree. It's not an error (any more) if module is simply not acquiring.