Add support for signed registers
So far all registers were treated as unsigned values. But some ROC
registers are signed and ROC calibration needed to do conversion
on it's own. Not all ROC firmware would use the same width for
register, making the conversion on client side more difficult.

The signed support was added as an optional parameter to createParam()
function, pointing to a class that does conversion to/from raw
value and basic boundary checking. So far we support 2s complement
signed conversion and sign and magnitude conversion. Others can be
added later if needed.

ROCs' InOffset is using sign and magnitude format, IntRelease is in
2s complement. All ROC temperatures are in 2s complement. Other
ROC registers are supposedly unsigned.

This code has been verified manually and using calibrationClient
before commiting.