Multi-event rejection
This is the initial code for multi-event rejection in BNL ROC calculation.
It tries to find multiple events by looking for number of local maxima
in the raw data. When more than one local maxima is found in each
dimension, event is rejected as multi-event as the more than one
event in the reading can corrupt the real event.

Pulse threshold had to be raised to 10,000,000 or 10% events would
be rejected. It's to be experimented what the correct value is and
to study the multi-event raw samples distribution. Improvements to
the multi-event rejection code are likely depending on findings.