bugfix: Allow resetting the previous pulse time of flight.
Because DSP-T is not properly tagging end-of-pulse-data packet,
nED is doing that to be complient with ADARA protocol. It does it
by looking at the absoulute time of every data packet and when time
change is detected, he injects an empty data packet to ADARA with
the EOP flag set.

When replaying the same data over and over again, the last packet
in raw data doesn't get EOPed until another data packet comes along.
This happens when data is replayed the next time, causing an empty
injected packet with a future time. Any logic to prevent that would
be prone to failure, so the solution is external Reset PV that clears
the time of the previous data packets.