Atomicity for a single outstanding CmdReq request
After CmdReq PV is processed and sends out a request to nED,
it automatically disables itself. This prevents further processing
until it is later re-enabled by CmdRsp when a response is received.
By using DISP field, writing to a disabled PV will result in CA
error, which any sane client should recognize as a failed attempt.