Allow plugins to exchange parameters
Plugin can send a parameter to another plugin. For now int32 params only.
The sending plugin connects to the remote plugin's asynInt32 interface
and writes new parameter value using the dedicated asyn reason. The
parameter name and sending plugin name are transmitted in a structure.
It's up to plugins to agree on the parameter naming convention.