Attempt to rewrite the CROC rejection code
The new approach is much simpler and with less parameters. It was written from
scratch based on the understanding the detector operation and different mappings.
The resulting code is much simpler, but this is before we've tested and verified
the code in the field.
There's a PV VerifyMode that selects which algorithm to use and defaults to
legacy, at least for now.