Add support for ADCROC firmware v1.5
This firmware changes the meaning of PmRatePeriod, previously
each bit was 8ms, now 16ms.  This allows the period that
we call "Minimum Repeat Time" in the OPI to be doubled, up
to a new range of 17.4 minutes.

This firmware also introduces some new registers related to
a Heat Pulser, which has not been tested.  The register definitions
were added anyway which is required for configuration reads and
writes to operate.

These changes were tested successfully on Corelli.