bugfix: Proper fix for PVA Metadata channel
This is a follow-up bugfix from Feb 5th where a workaround was made to
eliminate double userTags and timestamps. At that point flushing() data
queues was disabled as this seemed to cure but the fix did not make any
sense. This time the bugfix is complete, re-enabling flushing queues.

The attempt to make Metadata writes atomic, initial beginGroupPut()
was left out. Apparently caused un-expected issues.