Turns out ROC 5.7 *does* provide working Configured register
Previously the Configured and Discovered registers were not connected
to the hardware health PV. The misconception could be easily explained
by the fact that those two PVs never changed state when the module is
reset with a commend. But when power-cycled, at least Configured
register does reflect the proper state. This patch connects Configured
PV to the HwStatus.

This patch better explains missing/wrong functionality of ROC 5.7:
* sw reset doesn't properly clear all the registers
* Discovered register seems not connected