Allow plugins to connect to multiple sources
So far each plugin could have multiple childs but a single parent. With
this patch many parents are allowed. If proven to work as reliable and
as efficient as current setups, this patch would allow to reduce number
of plugins required on beamlines with more than one DSP. It would
reduce resources (now 3 pixel mapping plugins looking at the same file),
as well as clean up PV name space (removing the burden from python
clients which now have to know where the detector is connected to).

ADnED and SMS are both ready for the change.