Improved echo detection and removed low-count events.
Both changes were used in the 5 hour diamond run on 07/18 with
good results.
Echo detection is now aware of pulses and will reset the rejection
on every pulse change. Echoing is reducing the tail effect of pulses.
Low-counts rejection was disabled and it seems to reduce background
noise considerably. Maybe this can be made optional and called
efficiency boost.