bugfix: Index into flat field correction table could be out of bounds.
Surprisingly this hasn't been seen yet at BL-4B even after 5 months of
non-stop collecting data. It was only observed today when we changed
the position calculation to use a single channel to determine position
and that put the pixel right on the edge of the table, voila the
segfault. Luckily we have a debug build and it was easy to find the
problem. Need to deploy it ASAP.