Final port of dcomserver code.
This is the version that POWGEN team have verified to eliminate all the
ghosting and produce the same results as previously with dcomserver. It's
a cleaned up copy of dcomserver algorithm, with all the same parameters
and features. Some of the parameter are hard-coded and need to be made PVs.