Upgraded ProxyPlugin with down-ward communication.
Based on Bogdan's branch and work, the ProxyPlugin can now receive
command packets from the socket and relay them to OCC. The recv()
function was moved to BaseSocketPlugin.cpp for any class to benefit.
ProxyPlugin uses a thread to listen for incoming packets. Upgraded
Bogdan's work with nED helper functions for checking for packet
validity. Upgraded DB with autosave stuff similar to AdaraPlugin.