Add not-yet-tested support for FEM's new remote upgrade
Previous attempt to do FEM remote upgrades was never finalized, however
it did add some hooks. It was mostly relying on external state machine
or sequencer to drive sending all packets. This time it's implemented as
internal state machine for 2 reasons: avoid slight overhead of passing
information between nED core and external sequence, but most importantly
it can be paralelized without the need of one thread per module. In fact
there's no additional thread with this approach as all communication is
done as soon as response is received - it's processed in the receive
thread.

For this, internal response handler support was added to nED which can
be used for other purposes. It allows internal logic to send commands
independantly of those through PVs as long as they don't overlap. It's
expected other state machines are going to be implemented soon, like the
one for 'atomic' write config.