Working FEM remote upgrade support
After a successful first-time-ever remote upgrade of a FEM v3.9 module,
this code is commited as is. It implements a state machine to read MCS
file, split into small packets and send it over to FEM. Check upgrade
status every time and possibly wait for FEM to get ready.

Part to enable and disable remote upgrade support in FEM is currently
left out and needs to be operated manually.