FlatField: Position id change
The position id as received in data packet from ACPC is now called pixel
offset and is only kept internally. Detector id or position id is now
the identifier as known to scientists. This required changes to internal
book-keeping. There's a few more Veto PVs to distinguish different
rejection, with Position Config veto having alarm set at 1 - meaning
there's a conflict in configuration that needs to be resolved.