bugfix: Bad value written to 32bit register for LVDS modules.
As was spotted as AROC TimeVetoHigh and TimeVetoLow write/read mismatch
problem, it turned out a nED issue when 32bit register was spanning over
odd address. Since LVDS modules units are 16 bits but some of the
registers are 32 bits and because not all register land on even addresses,
nED has to do arithmetics. For almost 2 years now this bug has been
in code and very rarely exercised. In fact out of all tested modules
only LPSD ROC has 1 register TsyncDelay that is like that and in all
our cases the value of that register is 0.