Add support for programming sequence id
To be able to detect missed response from uprading module or to detect
missed command, module is reporting its sequence number that starts from
0 on init and increments by 1 for every received data packet. When
software is not sure whether the last packet was received, an empty
0x6F (remote upgrade) packet can be sent to retrieve the sequence id.
If sequence id did not change, same payload can be sent again.
A bunch of PVs were added to tweak all the thresholds.