bugfix: CmdDispatcher does not wait for consumer plugins
This issue has been revelead when switching to RHEL7. Somehow the
threading model or memory management were changed in RHEL7.

The problem was with the CmdDispatcher plugin sending command responses
to module plugins and not waiting for threaded plugins to complete,
instead deallocating the synchronization object. Usually modules are not
run in their own threads, it was a lucky misfortune to have them
mis-configured on MaNDI.

The fix is to wait for all plugins. On top there's now an assert in the
sync object that will fail more gracefully and precisly indicating the
problem shell we encounter similar problem in the future.