bugfix: AdaraPlugin race condition when sending RTDL and heart beat packets
Initially heart-beat was sent out when there were no packets sent to socket.
Later this was change on SMS request RTDL packets not to reset timer for
RTDL packets, so when there was no event data packets RTDL and heart-beat
could compete for socket. Since RTDL packet was split into two socket
writes, that allowed for a race condition with heart-beat write which got
injected into the output stream. Socket is opened in blocking mode that's
why the write()s are called from unlocked scope. Since write() is atomic
operation, the fix is to make RTDL packet a single write() operation.