Allow Discover plugin to automatically connect to parent plugins.
Getting rid of explicit Enable PV which was often left checked and the
plugin was parsing command responses all the time, while consuming CPU.
When a discover is triggered, plugin connects to configured parent
plugins, sends the message and wait 10 seconds for all responses to
come back. It then disconnects. This way it consumes no resources.