Make PluginMessage just a wrapper.
Previously a multi-inheritance was used was Das*PacketList message types.
This had a side effect that message and all its contents needed to be
copied when message was sent to other plugins. Instead a thin wrapper
was created holding a generic pointer for actual message data, with
convenient function to retrieve data casted to user liking. It's a
cleaner design and more efficient too.