Move Das1CompatibilityPlugin functionality to OCC.
DAS 1.0 and DAS 2.0 protocols are not completely disjoint - there's a
case when old DSP mirrors back the command which presents a case where
DAS 2.0 packet can no longer be uniquely distinguished from DAS 1.0
packet. Thus OCC needs a switch for DAS 1.0/2.0 packets and so does nED.
Rather than having multiple switches in different plugins,
Das1CompatibilityPlugin plugin was removed and its functionality moved
directly to the packets source or to the DasPacket implementation where
it belongs - the drawback is a single conditional
check for each packet but that should not harm performance a bit.