bugfix: Memory leak when scheduling no-response-callback.
There were two problems: the easy one was that BaseModulePlugin kept
the instance of the timer until next command was scheduled. This was not
really a problem but showed up in valgrind as lost memory since nobody
cleaned it up.

Second one was more severe. It kept all shared_ptr references of all
scheduled timers due to the circular referencing and which never dropped
usage counter to 0 and freed shared_ptr.