Add support for other detector's data formats.
Re-arranged code to be easily extended for other event formats, and add
support for BNL ROC raw format. Went crazy with C++ templates to achieve
the result but it turned out quite nice looking code which can be easily
extended for other formats. Better documenting functionality too.

Verified that the output is the same as before.