Improved BNL position and FlatField veto flagging.
Rejected events were wrongly flagged with veto in the position field,
which created un-existing detector id with all the rejected events
in it. Instead, flag pixelid with globally recognized veto flag and
invalidate x&y or corrected_x&corrected_y fields by putting -1 in it
rather than 0 which is in valid range. This way clients can detect
something was wrong during the calculation.