Make CmdReq PV set severity and status when it doesn't send command out.
Sending second command while waiting response from first command is illegal in nED.
Previously we disabled the CmdReq PV in database but that didn't communicate the
failure back to client and would lead to silent command drops (observed once with
TimingHandler which is bad). Instead the device support in nED now sets status and
severity which propagates to PV and client can check for it (TimingHandler is already
doing so).