bugfix: a segmentation fault caused by no-response cleanup handler
It was strange to see nED crash only when detectors got unplugged. Running
it through debugger it crashed twice in no response callback routine where
the internal structures were all corrupted. It wasn't obvious at first but
after running several scenarious in my head I concluded it must be the
wrong use of Timer and a potential race-condition when it is expiring.
There was a commit made in March 339dcdba that probably introduced this bug.
Ironically it was trying to fix another bug back then. Based on this
conclusion, a test patch was prepared and tested at TOPAZ for >24h
where before it was failing every few hours. No more crashes and the code
is much cleaner - hopefully no more bugs here.

The solution here is to avoid BasePlugin timer since it would get complicated
with locking the port driver. Instead we call Timer class directly and handle
all the locking ourselves. It's actually less code with an explicit lock(),
but I figured this would be okay. Derived plugins need to be aware of locking
when doing processing in own threads.