Removing BasePlugin::scheduleCallback() function
The scheduleCallback() function was at first introduced to simplify
invoking periodic or one-shot tasks from derived plugins and allowed the
plugins not to deal with port locking. Since then it has grown complex
and difficult to grasp. Corner cases were found which made derived
plugins to be carefull how to use this function, which defeats the
purpose of its existance. Removing the function and let plugins use
Timer class directly.