Clean-up BaseModulePlugin and introduce generic req/rsp handlers.
This is a move towards easier integration of register-based modules,
but it really is a simplification of a lot of bloated code covering many
weird communication exceptions in the generic class. It was originally
developed when adding module support graduadely but it became complicated
and hard to maintain. The new approch is hopefully much simpler with
a table of request/response handlers that modules can register. Defaults
are in place for all previous functionality so not too many changes in
derived classes were necessary. Check the Doxygen comment of
BaseModulePlugin for all details.