Move BaseModulePlugin to register based approach completely
The remaining functions were discover and version responses. The only
reason they stuck so long was the ModulesPlugin using them from outside.
Reading in the AROC IBC number caused me thinking about getting status
PVs in. And after figuring out a way to instantiate modules, it made all
the difference in how to talk to modules. BaseModulePlugin was simplified
A LOT, now all the functions are very generic and command agnostic. The
individual modules derivates were simplified too and are now mostly
shells for invoking version specific functions that create all PVs. The
default database had to be slightly changed too but again it's become
simpler. This patch concludes work on moving towards register based
modules - it makes it really easy to add those and cleans up the command
processing quite a bit.

"Perfection [in design] is achieved, not when there is nothing more to
add, but when there is nothing left to take away."
- Antoine de Saint-Exupéry