db: Double-locked PVs prevent accurate sequences
CmdRsp and CmdReq PVs are double-connected and when one can't process
until the other finishes. This does not work well for sequences.
A simple but not completely bullet-proof solution is to add small delay
between each command and hope that response is fully processed before
next command is sent out.