Clean up data packets sent to SMS
There's a suspicion that RtdlHeader structure didn't initialize
properly when constructed, so when RTLD data couldn't be found
in FIFO, the data packet sent to SMS would have garbage data in it.
Second part of the patch is to improve packet sequence numbers
and now we should have both what SMS calls dspSeq and packetSeq
implemented properly.