PVA channel optimization for high-throughput experiments
First I thought that the huge number of small updates is causing nED
and ADnED slow-downs. So the first attempt was to buffer data and only
send an update when timestamp change is detected. This should work even
for low rate experiements since we're receiving a 'timing' packet from DSP
for each pulse. However, the change didn't have significant impact on
performance.

The second step was to optimize memory allocation. It turns out PVAccess
memory allocation is very inefficient, but luckily we can use pre-allocated
buffer. We're using std::vector as a buffer, keep extending it if needed
but never de-allocate memory.

With both approaches, the performance is up quite a bit and nED is able to
sustain 21 mio ev/s as demonstrated at NOMAD. Based on CPU utilization
projection, it could do around 40 mio ev/s.